未使用コードを削除
- Server Action 経由のシミュレーション実行は廃止し、クライアントから直接 API Route を fetch する方式に統一済みだったが、残っていた `run-simulation-action.ts` の関数本体を削除
- `TransientConfigSnapshot` / `RunSimulationActionParams` (→ `SimulationRunRequest` に改名) の型定義を `shared/types/index.ts` に集約
- 旧「現行 vs 改善版 Judge」で使っていた `JUDGE_CRITERIA` / `JudgeCriterionKey` を削除（機能ごと廃止済み）
